move the flattening into the preset
